test: separate RecordingHapticExecutor from test code
